Fix circular import and reorganize prefixes
