+++ helpers/prefix_test.py
@@ -1,0 +1,2 @@
+# Präfix für Test-Tracker
+PREFIX_TEST = "TEST_"
