--- helpers/prefixes.py
@@ -1,3 +1,0 @@
-# Zusätzliche Präfixe für Marker
-PREFIX_TEST = "TEST_"
-PREFIX_RECOVERED = "RECOVERED_"
