+++ operators/tracking/solver.py
@@ -9,7 +9,7 @@ from ...helpers.utils import jump_to_frame_with_few_markers
 from ...helpers.prefix_new import PREFIX_NEW
 from ...helpers.prefix_track import PREFIX_TRACK
 from ...helpers.prefix_good import PREFIX_GOOD
-from ...helpers.prefixes import PREFIX_TEST, PREFIX_RECOVERED
+from ...helpers.prefix_test import PREFIX_TEST
 from ...helpers.select_track_tracks import select_track_tracks
 from ...helpers.select_new_tracks import select_new_tracks
 from ...helpers.feature_math import (
